backend/requirements.txt
flask
flask-cors
sqlalchemy
pyodbc
python-dotenv
openai
google-genai
gunicorn
waitress
